pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adaptive-model-routing-serving-optimizer"
version = "0.1.0"
description = "Intelligent model-serving gateway with dynamic routing across compression variants"
authors = [
    {name = "Alireza Shojaei", email = "alireza.shojaei@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "mlflow>=2.0.0",
    "ray[serve]>=2.0.0",
    "tritonclient[all]>=2.25.0",
    "onnx>=1.12.0",
    "onnxruntime>=1.12.0",
    "tensorrt>=8.5.0",
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "psutil>=5.9.0",
    "GPUtil>=1.4.0",
    "prometheus-client>=0.14.0",
    "requests>=2.28.0",
    "asyncio-mqtt>=0.11.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "jupyter>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true